llm_solutions/requirements.txt
# LLM Evaluation System Requirements

# Core async HTTP client
aiohttp>=3.8.0

# Data manipulation and analysis
pandas>=1.5.0

# Rate limiting and async utilities
asyncio-throttle>=1.0.0

# Optional: For better progress tracking
tqdm>=4.64.0

# For configuration management
python-dotenv>=0.19.0

# For testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# For improved JSON handling
orjson>=3.8.0

# For better logging
structlog>=22.0.0

# For API retries and backoff
tenacity>=8.0.0

# HuggingFace dependencies (optional - for local models)
# Uncomment these lines if you plan to use local HuggingFace models
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0